fix: renamed report base to selector base filter abstract component
