+++ apps/gauzy/src/app/pages/reports/amounts-owed-report/amounts-owed-report/amounts-owed-report.component.ts
@@ -12,7 +12,7 @@ import { pluck } from 'underscore';
 import { isEmpty } from '@gauzy/common-angular';
 import { Store } from './../../../../@core/services';
 import { TimesheetService } from './../../../../@shared/timesheet/timesheet.service';
-import { ReportBaseComponent } from './../../../../@shared/report/report-base/report-base.component';
+import { BaseSelectorFilterComponent } from './../../../../@shared/timesheet/gauzy-filters/base-selector-filter/base-selector-filter.component';
 import { IChartData } from './../../../../@shared/report/charts/line-chart/line-chart.component';
 import { ChartUtil } from './../../../../@shared/report/charts/line-chart/chart-utils';
 import { getAdjustDateRangeFutureAllowed } from './../../../../@theme/components/header/selectors/date-range-picker';
@@ -23,7 +23,7 @@ import { getAdjustDateRangeFutureAllowed } from './../../../../@theme/components
 	templateUrl: './amounts-owed-report.component.html',
 	styleUrls: ['./amounts-owed-report.component.scss']
 })
-export class AmountsOwedReportComponent extends ReportBaseComponent
+export class AmountsOwedReportComponent extends BaseSelectorFilterComponent
 	implements OnInit, AfterViewInit {
 
 	logRequest: IGetExpenseInput = this.request;
