+++ apps/gauzy/src/app/pages/reports/project-budgets-report/project-budgets-report/project-budgets-report.component.ts
@@ -19,15 +19,15 @@ import { TranslateService } from '@ngx-translate/core';
 import { isEmpty } from '@gauzy/common-angular';
 import { Store } from './../../../../@core/services';
 import { TimesheetService } from './../../../../@shared/timesheet/timesheet.service';
-import { ReportBaseComponent } from './../../../../@shared/report/report-base/report-base.component';
+import { BaseSelectorFilterComponent } from './../../../../@shared/timesheet/gauzy-filters/base-selector-filter/base-selector-filter.component';
 
-@UntilDestroy()
+@UntilDestroy({ checkProperties: true })
 @Component({
 	selector: 'ga-project-budgets-report',
 	templateUrl: './project-budgets-report.component.html',
 	styleUrls: ['./project-budgets-report.component.scss']
 })
-export class ProjectBudgetsReportComponent extends ReportBaseComponent 
+export class ProjectBudgetsReportComponent extends BaseSelectorFilterComponent 
 	implements OnInit, AfterViewInit {
 		
 	logRequest: IGetPaymentInput = this.request;
